added rough side navigation bar
