+++ frontend/app/components/disruption.vue
